+++ sensors/core/mag.go
@@ -7,7 +7,7 @@ import (
 // MagReading represents 3 axis magnetometer readings in micro-Teslas
 type MagReading struct {
 	Timestamp time.Time
-	XuT       float64
-	YuT       float64
-	ZuT       float64
+	XuT       float32
+	YuT       float32
+	ZuT       float32
 }
